requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==4.0.1
flask-sock==0.7.0
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.0.1
pillow==10.4.0
simple-websocket==1.0.0
Werkzeug==3.0.3
wsproto==1.2.0
